divided it into translation units and added a namespase
