+++ number.txt
@@ -1,0 +1,2 @@
+2 3
+2 5 6 4 2 7
